benet: fix build error on 32bit arch
Patchwork-id: 44804
O-Subject: [RHEL6.3 PATCH 09/62] benet: fix build error on 32bit arch
Bugzilla: 738041
RH-Acked-by: David S. Miller <davem@redhat.com>

BZ: #738041

Upstream commit:
commit 68c3e5a7b91513010d2536e4bcd7bdd54d0e6acf
Author: Eric Dumazet <eric.dumazet@gmail.com>
Date:   Tue Aug 9 06:23:07 2011 +0000

    benet: fix build error on 32bit arch

    Error comes from commit ac124ff973e27802797
    (be2net: cleanup and refactor stats code)

    ERROR: "__udivdi3" [drivers/net/benet/be2net.ko] undefined!

+++ drivers/net/benet/be_main.c
@@ -932,7 +932,7 @@ static void be_rx_eqd_update(struct be_adapter *adapter, struct be_rx_obj *rxo)
 	if (delta < HZ)
 		return;
 
-	stats->rx_pps = (stats->rx_pkts - stats->rx_pkts_prev) / (delta / HZ);
+	stats->rx_pps = (unsigned long)(stats->rx_pkts - stats->rx_pkts_prev) / (delta / HZ);
 	stats->rx_pkts_prev = stats->rx_pkts;
 	stats->rx_jiffies = now;
 	eqd = stats->rx_pps / 110000;
